Move Time Travel Debugging to separate module
- Create flowdux-timetravel module as optional extension
- Remove timetravel code and kotlinx-datetime from core flowdux
- Update README with separate module installation instructions
- Keep core library lightweight without debugging dependencies

Usage:
  implementation("com.github.chibimoons.flowdux:flowdux:x.x.x")
  implementation("com.github.chibimoons.flowdux:flowdux-timetravel:x.x.x")
