+++ flowdux-timetravel/src/commonMain/kotlin/io/flowdux/timetravel/StateSnapshot.kt
